import release 26.2 from source262zip

+++ primenet.c
@@ -969,9 +969,9 @@ int format_args (char* args, short operation, void* pkt)
 		/* Server does not like a pcercent complete of 100%. */
 		/* Just in case caller passes that value in, convert it */
 		sprintf (p, "&c=%lu&p=%.4f&d=%lu&e=%lu",
-			 z->cpu_num,
+			 (unsigned long) z->cpu_num,
 			 z->pct_complete < 99.9999 ? z->pct_complete : 99.9999,
-			 z->next_update, z->end_date);
+			 (unsigned long) z->next_update, (unsigned long) z->end_date);
 		p += strlen (p);
 		if (z->fftlen) {
 			sprintf (p, "&fftlen=%d", z->fftlen);
